pyproject.toml
[tool.poetry]
name = "strava-to-postgres"
version = "0.1.0"
description = ""
authors = ["John J Czaplewski"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
polyline = "^1.4.0"
python-dotenv = "^0.19.0"
psycopg = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
